refactor(memory): split reflection into package
