+++ tests/unit/test_reflection_refine_apply.py
@@ -38,7 +38,7 @@ def _install(tmpdir):
     with patch("memory.event_log.get_config_manager", return_value=cm), \
          patch("memory.persona.get_config_manager", return_value=cm), \
          patch("memory.facts.get_config_manager", return_value=cm), \
-         patch("memory.reflection.get_config_manager", return_value=cm):
+         patch("memory.reflection.manager.get_config_manager", return_value=cm):
         event_log = EventLog()
         fact_store = FactStore()
         pm = PersonaManager(event_log=event_log)
